// src/enums/tweet-fields.ts
// -----------------------------------------------------------------------------------------
// #region Enums
// -----------------------------------------------------------------------------------------

enum TweetFields {
    Attachments = "attachments",
    AuthorId = "author_id",
    ContextAnnotations = "context_annotations",
    ConversationId = "conversation_id",
    CreatedAt = "created_at",
    Entities = "entities",
    Geo = "geo",
    Id = "id",
    InReplyToUserId = "in_reply_to_user_id",
    Lang = "lang",
    NonPublicMetrics = "non_public_metrics",
    PublicMetrics = "public_metrics",
    OrganicMetrics = "organic_metrics",
    PromotedMetrics = "promoted_metrics",
    PossiblySensitive = "possibly_sensitive",
    ReferencedTweets = "referenced_tweets",
    Source = "source",
    Text = "text",
    Withheld = "withheld",
}

// #endregion Enums

// -----------------------------------------------------------------------------------------
// #region Exports
// -----------------------------------------------------------------------------------------

export { TweetFields };

// #endregion Exports