// src/enums/tweet-expansions.ts
// -----------------------------------------------------------------------------------------
// #region Enums
// -----------------------------------------------------------------------------------------

enum TweetExpansions {
    AttachmentsPollIds = "attachments.poll_ids",
    AttachmentsMediaKeys = "attachments.media_keys",
    AuthorId = "author_id",
    EntitiesMentionsUsername = "entities.mentions.username",
    GeoPlaceId = "geo.place_id",
    InReplyToUserId = "in_reply_to_user_id",
    ReferencedTweetsIds = "referenced_tweets.id",
    ReferencedTweetsIdAuthorId = "referenced_tweets.id.author_id",
}

// #endregion Enums

// -----------------------------------------------------------------------------------------
// #region Exports
// -----------------------------------------------------------------------------------------

export { TweetExpansions };

// #endregion Exports