// src/enums/user-fields.ts
// -----------------------------------------------------------------------------------------
// #region Enums
// -----------------------------------------------------------------------------------------

enum UserFields {
    CreatedAt = "created_at",
    Description = "description",
    Entities = "entities",
    Id = "id",
    Location = "location",
    Name = "name",
    PinnedTweetId = "pinned_tweet_id",
    ProfileImageUrl = "profile_image_url",
    Protected = "protected",
    PublicMetrics = "public_metrics",
    Url = "url",
    Username = "username",
    Verified = "verified",
    Withheld = "withheld",
}

// #endregion Enums

// -----------------------------------------------------------------------------------------
// #region Exports
// -----------------------------------------------------------------------------------------

export { UserFields };

// #endregion Exports